+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/config/basic.php
@@ -5,9 +5,7 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 use App\BarService;
 
 return function (ContainerConfigurator $c) {
-
     $s = $c->services();
     $s->set(BarService::class)
         ->args(array(inline('FooClass')));
-
 };