+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/containers/container9.php
@@ -4,6 +4,7 @@ require_once __DIR__.'/../includes/classes.php';
 require_once __DIR__.'/../includes/foo.php';
 
 use Symfony\Component\DependencyInjection\Argument\IteratorArgument;
+use Symfony\Component\DependencyInjection\Argument\TaggedIteratorArgument;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
@@ -159,6 +160,16 @@ $container
 $container
     ->register('lazy_context_ignore_invalid_ref', 'LazyContext')
     ->setArguments(array(new IteratorArgument(array(new Reference('foo.baz'), new Reference('invalid', ContainerInterface::IGNORE_ON_INVALID_REFERENCE))), new IteratorArgument(array())))
+    ->setPublic(true)
+;
+$container
+    ->register('tagged_iterator_foo', 'Bar')
+    ->addTag('foo')
+    ->setPublic(false)
+;
+$container
+    ->register('tagged_iterator', 'Bar')
+    ->addArgument(new TaggedIteratorArgument('foo'))
     ->setPublic(true)
 ;